move parser to utils
