+++ src/client/createQuizApp/components/pages/CQEdit/_CQEditStyles.scss
@@ -98,6 +98,7 @@
 
     &__import-button {
         position: relative;
+        float: right;
     }
 
     &__footer {
