+++ src/client/createQuizApp/components/utils/CQQuestionUploader.js
@@ -1,10 +1,9 @@
 /* @flow */
 import React from 'react';
 
-import {
-    QuizStore
-} from './../../stores';
+import { importFileParser } from './../../utils';
 
+import type { Question } from './../../../../types';
 
 class CQQuestionUploader extends React.Component {
 
@@ -22,7 +21,7 @@ class CQQuestionUploader extends React.Component {
                 var data = upload.target.result;
                 this.setState({ data });
                 this.props.onQuestionRawData(data);
-                this.props.onQuestionData(this.processData(data));
+                this.props.onQuestionData(importFileParser(data, this.props.format, this.props.quizId));
             };
 
             var dataFile = ev.target.files[0];
@@ -37,49 +36,7 @@ class CQQuestionUploader extends React.Component {
     }
 
 
-    processData (data: string): Question {
-        if (this.props.format == "doodlemath") {
-            return this.processDoodleMath(data);
-        }
 
-    }
-
-    processDoodleMath (data: string): Question {
-        let quizId = this.props.quizId;
-        let lines = data.trim().split('\n');
-        let columns = lines.splice(0, 1)[0].split('\t');
-        let processedData = lines.map(function (line, index) {
-            let q = line.split('\t');
-            let question = {};
-            let newQuestion = QuizStore.getQuestion(quizId, index);
-            for (let i = 0; i < columns.length; i++) {
-                question[columns[i]] = q[i];
-            }
-            question.options = [question.A, question.B, question.C, question.D];
-            newQuestion.question = question.Question;
-            switch (question.QuestionType) {
-                case "T/F":
-                    newQuestion.answer = `boolean://${question.options[0].toLowerCase()}`;
-                    break;
-                case "Multiple choice":
-                    newQuestion.answer = question.options.splice(0, 1)[0];
-                    newQuestion.alternatives = question.options;
-                    break;
-                case "Text":
-                    newQuestion.answer = `freetext://${question.options[0]}`;
-                    break;
-                case "Sorting6":
-                case "Sorting":
-                    newQuestion.answer = `sorting://${question.options.filter(o => o).join(":")}`;
-                    break;
-                case "Linking":
-                    newQuestion.answer = `linking://${question.options.filter(o => o).join(":")}`;
-                    break;
-            }
-            return newQuestion;
-        });
-        return processedData;
-    }
 
     render () : any {
         return (
